Load bodies directly as base64 rather than via XHR
The XHR method we'd been using works great if you're loading this over
HTTP, but it doesn't work when loading directly from the local
filesystem.

Since IRE is asking that we avoid relying on the classroom's internet
connection for our hands-on materials, let's load the files this way so
browsing to `index.html` at a `file://` path still works as expected.

Besides, the bodies' total size (before base64 conversion) was an order
of magnitude smaller than that of the rest of the JS in here. It really
isn't a big deal.